pyproject.toml
[project]
name = "aipractice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "agents>=1.4.0",
    "automium-web>=0.1.1",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.14",
    "python-dotenv>=1.2.2",
]
